requirements.txt
google-adk
streamlit
pandas
scikit-learn
python-dotenv
google-generativeai
nest_asyncio
